% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.scc.R
\name{extract.scc}
\alias{extract.scc}
\title{extract.scc}
\usage{
extract.scc(model, neg = "af")
}
\arguments{
\item{model}{a brms object}

\item{neg}{a character}
}
\value{
a dataframe of posterior distributions
}
\description{
function for extracting posterior distributions generated by a simple intercept model tailored for single color controls
}
\examples{
\dontrun{
}
}
