% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{geo.mean}
\alias{geo.mean}
\title{geo.mean}
\usage{
geo.mean(x, na.rm = T)
}
\arguments{
\item{x}{a vector}
}
\description{
calculates the geometric mean a robust metric of the central tendency of a population. applicable to values on a log-normal scale is robust to outliers will only ever be at most equal to the arithmetic mean
}
