% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.density.R
\name{plot.density}
\alias{plot.density}
\title{plot.density}
\usage{
\method{plot}{density}(data, x = fsc, y = ssc)
}
\arguments{
\item{data}{a dataframe of flow cytometry data}

\item{x}{an object; the channel to plot on the x axis; defaults to fsc (forward scatter)}

\item{y}{an object; the channel to plot on the y axis; defaults to ssc (side scatter)}

\item{batch}{a logical of whether to batch process or not. if TRUE, adds additional facets by sample id. defaults to FALSE.}
}
\value{
a ggplot object
}
\description{
a wrapper for making a ggplot object of the density of cells on two axes. creates a density column that estimates distance of each observation from the center of the channel. the product of density on both channels defines the density of cells.
}
\examples{
\donrun{plot.density(data = data)}
}
