% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.wave.R
\name{plot.wave}
\alias{plot.wave}
\title{plot.wave}
\usage{
\method{plot}{wave}(data)
}
\arguments{
\item{data}{a dataframe of flow cytometry data}
}
\description{
wrapper for making a ggplot object that plots the fluorescent intensities in each channel of a dataframe as density plots (waves).
}
\examples{
\dontrun{
}
}
