% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.fcs.R
\name{process.fcs}
\alias{process.fcs}
\title{process.fcs}
\usage{
process.fcs(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a dataframe
}
\description{
a pipeline for setting a standard processing protocol on flow cytometry data. log transforms all values, filters offscale events, then gates to most "average" cells in the population - within 1 s.d. of the median in all channels.
}
\examples{
\dontrun{
}
}
