% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{re.code}
\alias{re.code}
\title{re.code}
\usage{
re.code(x, mapping)
}
\arguments{
\item{x}{a character or factor vector}

\item{mapping}{a vector of matched channel names to label names}
}
\description{
function for re-labelling channels with new parameter names
}
