% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.fcs.R
\name{read.fcs}
\alias{read.fcs}
\title{read.fcs}
\usage{
read.fcs(path)
}
\arguments{
\item{file}{a character string that indicates a path to a *.fcs file. If a directory is given, all *.fcs files in the directory will be read.}
}
\value{
a dataframe that contains the filename, ID, channel name and corresponding values
}
\description{
simple wrapper that reads a *.fcs to a dataframe
}
\examples{
\dontrun{
}
}
