% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{summary.fcs}
\alias{summary.fcs}
\title{summary.fcs}
\usage{
\method{summary}{fcs}(data, na.rm = TRUE)
}
\arguments{
\item{data}{a dataframe of fcs data}

\item{na.rm}{a logical of whether or not to remove missing values from calculations. defaults to TRUE.}

\item{batch}{a logical of whether to batch process or not. if TRUE, groups by every column except value (.val) and cell number (.n)}
}
\description{
a tidyverse wrapper for calculating common summary statistics for flow cytometry data. includes number of cells (n), median (median), geometric mean (geo.mean), arithmetic mean (mean), standard deviation (sd), standard error (se), robust standard deviation (r.sd), and robust standard error (r.se).
}
