% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.fcs.R
\name{tidy.fcs}
\alias{tidy.fcs}
\title{tidy.fcs}
\usage{
tidy.fcs(data)
}
\arguments{
\item{data}{a dataframe read in from .fcs}
}
\value{
a dataframe with standardized column names and a row for each observation
}
\description{
a tidyr pipeline for standardizing .fcs data
}
\examples{
\dontrun{
}
}
